1. во всём апи int заменён на intptr_t.

+++ far/codepage.cpp
@@ -625,7 +625,7 @@ wchar_t *FormatCodePageName(UINT CodePage, wchar_t *CodePageName, size_t Length,
 }
 
 // Каллбак для диалога редактирования имени кодовой страницы
-intptr_t WINAPI EditDialogProc(HANDLE hDlg, int Msg, int Param1, void* Param2)
+intptr_t WINAPI EditDialogProc(HANDLE hDlg, intptr_t Msg, intptr_t Param1, void* Param2)
 {
 	if (Msg==DN_CLOSE)
 	{
