+++ far/filefilterparams.cpp
@@ -656,7 +656,7 @@ void FilterDlgRelativeDateItemsUpdate(HANDLE hDlg, bool bClear)
 	SendDlgMessage(hDlg,DM_ENABLEREDRAW,TRUE,0);
 }
 
-intptr_t WINAPI FileFilterConfigDlgProc(HANDLE hDlg,int Msg,int Param1,void* Param2)
+intptr_t WINAPI FileFilterConfigDlgProc(HANDLE hDlg,intptr_t Msg,intptr_t Param1,void* Param2)
 {
 	switch (Msg)
 	{
