+++ far/usermenu.cpp
@@ -868,7 +868,7 @@ enum EditMenuItems
 	EM_BUTTON_CANCEL,
 };
 
-intptr_t WINAPI EditMenuDlgProc(HANDLE hDlg,int Msg,int Param1,void* Param2)
+intptr_t WINAPI EditMenuDlgProc(HANDLE hDlg,intptr_t Msg,intptr_t Param1,void* Param2)
 {
 #if defined(PROJECT_DI_MEMOEDIT)
 	Dialog* Dlg=(Dialog*)hDlg;
